sdm: make sure the FOD pressed layer is correctly detected
This behaviour can be enabled or disabled (at runtime or at build time) by setting
the property debug.hwc.use_fod_index to 1 or 0, respectively.

Allocating HWC for Virtual Displays leads to the configuration of the
framebuffer target as GPU target composition type.

While the GPU target layer is at the top zorder (which is most of the time),
the flags and other attributes of the layers are the same as the framebuffer target layer
once reached SetupAtomic in DRM, making the FOD layer to not be detected correctly.
To solve this, add a second way of detecting the FOD layer: its index.

If the FOD layer is found in the HWC, a flag in the layer stack
will be set with its index, which correctly reaches SetupAtomic.

Once reached SetupAtomic, a check will be made for detecting the FOD flag.
If the FOD flag exists in some of ther layers, then that will be the method for detecting the layer.
Otherwise, the index will be used for detecting the layer, if the property is set to 1.

For getting the FOD layer index, discard all the non-hw layers below the FOD layer
so it is correctly detected at SetupAtomic.
Also discard the GPU target layer from the index (if existing).

Signed-off-by: daniml3 <danimoral1001@gmail.com>
Change-Id: Icc2233909218a5569a41065bb69e8d22c99497af