sdm: detect the FOD layer by using the layer index
Allocating HWC for Virtual Displays leads to the configuration of the framebuffer target as GPU target composition type.
For some reason, while the GPU target layer is at the top zorder (which is most of the time), the flags and other attributes of the layer are invalid once reached SetupAtomic in DRM.

To solve this, instead of setting a bit on the layer flags, set the FOD layer index in the layer stack, which reaches correctly SetupAtomic in DRM.
For getting the FOD layer index, discard all the non-hw layers below the FOD layer so it is correctly detected at SetupAtomic. Also discard the GPU target layer from the index (if existing).

Signed-off-by: daniml3 <danimoral1001@gmail.com>
Change-Id: Icc2233909218a5569a41065bb69e8d22c99497af