sdm: make sure the FOD pressed layer is correctly detected
This behaviour can be enabled or disabled (at runtime or at build time)
by setting the property debug.hwc.use_fod_index to 1 or 0 respectively.

Allocating HWC for Virtual Displays leads to the configuration of the
framebuffer target as GPU target composition type:

While the GPU target layer is at the top zorder (which is most of
the time), the flags and other attributes of the layers are the same
as the framebuffer target once reached SetupAtomic in DRM, making the
FOD layer to not be correctly detected.
To solve this, add a second way of detecting the FOD layer: its index.

If the FOD layer is found in the HWC, a flag in the layer stack
will be set with its index, which correctly reaches SetupAtomic.

Once reached SetupAtomic, the layer will be marked as FOD layer if
the layer has the FOD pressed flag or if its index matches the FOD layer
index in the layer stack, and the property is set to 1.

For getting the FOD layer index, discard all the non-hw layers below
the FOD layer so it is correctly detected at SetupAtomic.
Also discard the GPU target layer from the index (if existing).

test: start a VDS with HWC allocated (WFD), stop the session,
make sure that the FOD pressed layer is correctly marked

Signed-off-by: daniml3 <danimoral1001@gmail.com>
Change-Id: Icc2233909218a5569a41065bb69e8d22c99497af